+++ app/components/auth/AuthButton.tsx
@@ -1,40 +1,21 @@
 import React, { useEffect, useState } from 'react'
 
 import { useInternetIdentity } from './InternetIdentity'
-import { log } from '@/app/util/log'
 import { Button } from '../library/Button'
-import { CenterSpinner } from '../Spinner'
-import { queryUserController } from '@/app/client/user'
 import { useNavigate } from 'react-router-dom'
 
 export const AuthButton = () => {
-  const { authenticate, signout, isAuthenticated, identity, controllerId, setControllerId } = useInternetIdentity()
-  const [isLoading, setLoading] = useState(false)
+  const { authenticate, signout, isAuthenticated, controllerId, setControllerId } = useInternetIdentity()
   const navigate = useNavigate()
   const [isAutoRedirect, setIsAutoRedirect] = useState(false)
 
   useEffect(() => {
-    if (isAuthenticated && identity) {
-      const loadControllerId = async () => {
-        try {
-          setLoading(true)
-          const myControllerId = await queryUserController(identity)
-          setControllerId(myControllerId)
-
-          if (isAutoRedirect && myControllerId) {
-            navigate(`/ai/${myControllerId}`)
-          }
-        } catch (error) {
-          log.error('Error in loadControllerId', error as Error)
-        } finally {
-          setLoading(false)
-        }
-      }
-
-      loadControllerId()
+    if (isAutoRedirect && isAuthenticated && controllerId) {
+      navigate(`/ai/${controllerId}`)
     }
+
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [identity, isAuthenticated, navigate])
+  }, [controllerId, isAuthenticated, navigate])
 
   const processSignout = async () => {
     setControllerId(null)
@@ -48,12 +29,9 @@ export const AuthButton = () => {
 
   return (
     <>
-      {isLoading && <CenterSpinner />}
-      {!isLoading && (
-        <Button outline onClick={isAuthenticated ? processSignout : processAuthenticate}>
-          {isAuthenticated ? 'Logout' : 'Sign In'}
-        </Button>
-      )}
+      <Button outline onClick={isAuthenticated ? processSignout : processAuthenticate}>
+        {isAuthenticated ? 'Logout' : 'Sign In'}
+      </Button>
       {controllerId && <Button onClick={() => navigate(`/ai/${controllerId}`)}>Open ArcMind AI</Button>}
     </>
   )
